Refactor WorkerTagPicker to component
