+++ examples/deploy/aws-ecs-terraform/windmill_worker_native.tf
@@ -26,7 +26,7 @@ resource "aws_ecs_task_definition" "windmill_cluster_windmill_native_worker_td" 
         value = "true"
       }, {
         name  = "DATABASE_URL"
-        value = "postgres://${aws_db_instance.windmill_cluster_rds.username}:${aws_db_instance.windmill_cluster_rds.password}@${aws_db_instance.windmill_cluster_rds.endpoint}/${aws_db_instance.windmill_cluster_rds.db_name}"
+        value = "postgres://${aws_db_instance.windmill_cluster_rds.username}:${aws_db_instance.windmill_cluster_rds.password}@${aws_db_instance.windmill_cluster_rds.endpoint}/${aws_db_instance.windmill_cluster_rds.db_name}?sslmode=disable"
         }, {
         name  = "MODE"
         value = "worker"
